// packages/dashboard/src/http/types/ListProcesses.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ListProcessesQuery = {
    /**
     * @minLength 1
     * @maxLength 9007199254740991
     * @default 1
     * @type integer | undefined
    */
    page?: number;
    /**
     * @minLength 1
     * @maxLength 100
     * @default 20
     * @type integer | undefined
    */
    pageSize?: number;
};

export type ListProcessesStatus200 = {
    data: {
        id: string;
        caseNumber: string | null;
        clientId: string;
        plaintiff: string;
        defendant: string;
        notes: string | null;
        /**
         * @description
         * Format: `date-time`
         * @type string
        */
        createdAt: string;
        /**
         * @description
         * Format: `date-time`
         * @type string
        */
        updatedAt: string;
        client: {
            id: string;
            fullName: string;
        };
        user: {
            id: string;
            name: string;
        };
    }[];
    pagination: {
        page: number;
        pageSize: number;
        total: number;
        pageCount: number;
    };
};

export type ListProcessesOptions = {
    body?: never;
    path?: never;
    query?: ListProcessesQuery;
    headers?: never;
};

export type ListProcessesResponses = {
    "200": ListProcessesStatus200;
};

/**
 * @description Union of all possible responses
*/
export type ListProcessesResponse = ListProcessesStatus200;